# lib/api-zod/src/generated/types/compressRequest.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Auriga Audio Engine API
 * OpenAPI spec version: 0.1.0
 */
import type { EnvironmentObservation } from './environmentObservation';

export interface CompressRequest {
  observations: EnvironmentObservation[];
  maxWords?: number;
}