# lib/api-zod/src/generated/types/researchDebtEntry.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Auriga Audio Engine API
 * OpenAPI spec version: 0.1.0
 */
import type { ResearchDebtEntryCategory } from './researchDebtEntryCategory';
import type { ResearchDebtEntrySeverity } from './researchDebtEntrySeverity';

export interface ResearchDebtEntry {
  id: string;
  category: ResearchDebtEntryCategory;
  title: string;
  description: string;
  severity: ResearchDebtEntrySeverity;
  impact: string;
  assumptions: string[];
  failureModes: string[];
  createdAt: Date;
}