# lib/api-zod/src/generated/types/validationRequest.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Auriga Audio Engine API
 * OpenAPI spec version: 0.1.0
 */
import type { ValidationRequestProcedure } from './validationRequestProcedure';

export interface ValidationRequest {
  procedure: ValidationRequestProcedure;
  messageIds?: string[];
}